add backend services to fetch only doctors

+++ Hospital-management-system/src/main/java/CO225/Hospitalmanagementsystem/model/UserInfoModel.java
@@ -14,7 +14,7 @@ public class UserInfoModel {
     private String email;
     private String roles;
     private String nic;
-    private int wardNo;
+    private long wardNo;
     private String birthDate;
